+++ packages/@expo/image-utils/CHANGELOG.md
@@ -10,6 +10,10 @@
 
 ### 💡 Others
 
+## 0.8.16 — 2026-08-17
+
+_This version does not introduce any user-facing changes._
+
 ## 0.8.15 — 2026-07-15
 
 _This version does not introduce any user-facing changes._